fix: module search path
By default, Python adds the directory containing the script you're running to the beginning of this search path. This allows Python to import modules located in the same directory or subdirectories.

However, Python's import system doesn't automatically add parent directories to the module search path. This means that if you're running a script from a subdirectory and you want to import modules from parent directories, you need to either navigate to the parent directory before running the script or use absolute imports.